// clients/javascript/lib/gen_types/GetUserFreeBusyAPIResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { EventInstance } from "./EventInstance";

/**
 * API response for getting user free/busy
 */
export type GetUserFreeBusyAPIResponse = { 
/**
 * List of busy events
 */
busy: Array<EventInstance>, 
/**
 * UUID of the user
 */
userId: string, };