// clients/javascript/lib/gen_types/MultipleFreeBusyRequestBody.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ID } from "./ID";

/**
 * Request body for getting multiple free/busy
 */
export type MultipleFreeBusyRequestBody = { 
/**
 * List of user UUIDs to query
 */
userIds: Array<ID>, 
/**
 * Start time for the query (UTC)
 */
startTime: Date, 
/**
 * End time for the query (UTC)
 */
endTime: Date, };